Renamed and moved ScrollWrapper.js
