+++ client/src/App.js
@@ -52,7 +52,7 @@ import { stateChange } from "./helpers/Redirect";
 import { actions as appActions } from "./reducers/lab1/AppReducer";
 import useMainStateContext from "./reducers/MainContext";
 import { Spinner } from "reactstrap";
-import ScrollWrapper from "./use-hooks/scrollWrapper";
+import ScrollWrapper from "./helpers/ScrollWrapper";
 
 const LabWindow = lazy(
   () => import("./components/all-components/Lab/LabWindow"),
